Fix IRawElementProviderSimple.GetPropertyValue (#9732)
Our interop handle structs (such as HWND) cannot be marshalled directly and will fail "silently" on
callbacks (they will throw but the marshaller will convert that to an HRESULT that we won't see unless first-chance exceptions are on when we're debugging).

+++ src/System.Windows.Forms/src/System/Windows/Forms/Control.ControlAccessibleObject.cs
@@ -496,7 +496,7 @@ public partial class Control
                 UiaCore.UIA.LiveSettingPropertyId => Owner is IAutomationLiveRegion owner
                     ? owner.LiveSetting
                     : base.GetPropertyValue(propertyID),
-                UiaCore.UIA.NativeWindowHandlePropertyId => Owner?.InternalHandle ?? HWND.Null,
+                UiaCore.UIA.NativeWindowHandlePropertyId => (nint)(Owner?.InternalHandle ?? HWND.Null),
                 _ => base.GetPropertyValue(propertyID)
             };
